+++ core/balance.go
@@ -14,7 +14,6 @@ var (
 type Balance struct {
 	Address common.Address `json:"address" yaml:"address"`
 	Balance uint64         `json:"balance" yaml:"balance"`
-	Stake   uint64         `json:"stake" yaml:"stake"`
 	Nonce   uint64         `json:"nonce" yaml:"nonce"`
 	Symbol  string         `json:"symbol" yaml:"symbol"`
 	Units   string         `json:"units" yaml:"units"`